Track winners by name and report to players
Reports the last 10 winners to players.

Also, refactored so that ws.js is dealing with "connections," which are
sockets wrapped with other data associated with that socket (so far just
names).

Also, clients automatically call "exitGame" when the game ends, so that
the player immediately has the list of games they can join again.