Server and client disagreed on game height
The server's model of the game had it at 600 pixels tall, but the
client's model (and what was displayed to the user) was 500 pixels tall.
This led to the wrong winners being reported, when the clients thought a
collision had happened but the server had no idea.

Fixed this by creating a file common to both the client and the server,
settings.js, that tells them both the game dimensions, so that they both
get the dimensions from the same source, guaranteeing that they'll be
working with the same game area.